Create statements list with adapter and swip refresh layout

+++ BankApp/app/src/main/java/br/com/santander/android/bank/statement/domain/model/Statement.kt
@@ -15,5 +15,5 @@ data class Statement(
     val date: Date,
 
     @SerializedName("value")
-    val value: Double
+    val value: Float
 ): Serializable
